requirements-txt.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pillow==10.0.0
numpy==1.24.3
python-dotenv==1.0.0
pydantic==2.5.2
opencv-python-headless==4.8.1.78
torch==2.0.1
torchvision==0.15.2
transformers==4.35.0
gunicorn==21.2.0
